requirements.txt
aiogram==3.0.0b7
requests==2.28.1
python-dotenv==1.0.0

# Зависимости для тестирования
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-mock==3.10.0
aresponses==3.0.0
pytest-cov==4.1.0

